requirements.txt
fastapi
funasr
modelscope
numpy<=1.26.4
pydantic
pydantic_settings
soundfile
starlette
torch<=2.3
torchaudio
uvicorn
addict
datasets
pillow
simplejson
sortedcontainers
websockets
langchain-openai-0.2.0
langchain-0.3.0
pyaudio-0.2.14
transformers
langchain-community